// src/app/open-api/model/kundeResponse.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AdresseResponse } from './adresseResponse';
import { EmaResponse } from './emaResponse';
import { KontaktResponse } from './kontaktResponse';
import { ObjektIdentResponse } from './objektIdentResponse';
import { VideoResponse } from './videoResponse';
import { WartungResponse } from './wartungResponse';

export interface KundeResponse { 
    id?: number;
    kundenname?: string;
    videos?: Array<VideoResponse>;
    emas?: Array<EmaResponse>;
    adressen?: Array<AdresseResponse>;
    kontakte?: Array<KontaktResponse>;
    wartungen?: Array<WartungResponse>;
    objektidents?: Array<ObjektIdentResponse>;
}